Time: 0 ms (100%), Space: 7.1 MB (62.23%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     bool judgeSquareSum(int c) {
-        int x = 0, y = (int)sqrt(c);
+        long long x = 0, y = (long long)sqrt(c);
         while (x <= y) {
-            int k = x * x + y * y;
+            long long k = x * x + y * y;
             if (k == c) return true;
             else if (k > c) y--;
             else x++;
